chore: useGroupFilterInput refactor
